requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
attrs==24.2.0
bcrypt==4.2.1
Brotli==1.0.9
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.31.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
einops==0.8.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gradio==4.44.0
gradio_client==1.3.0
gradio_image_prompter==0.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.7
imageio==2.9.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl_fft
mkl_random
mkl-service
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.14.2
networkx==3.4.2
ninja==1.11.1.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
pillow==10.4.0
pip==24.0
pluggy==1.5.0
portalocker==3.0.0
proglog==0.1.9
propcache==0.2.0
psutil==5.9.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.0
safehttpx==0.1.1
scipy==1.14.1
semantic-version==2.10.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spaces==0.30.4
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
tomlkit==0.12.0
torch
torchaudio
torchvision
tqdm==4.67.1
triton==3.1.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.32.1
websockets==11.0.3
wheel==0.43.0
yarl==1.18.0
ultralytics
